+++ modules/apps/commerce/commerce-order-rule-web/src/main/resources/META-INF/resources/init.jsp
@@ -33,8 +33,8 @@ page import="com.liferay.commerce.order.rule.exception.COREntryExpirationDateExc
 page import="com.liferay.commerce.order.rule.exception.NoSuchCOREntryException" %><%@
 page import="com.liferay.commerce.order.rule.model.COREntry" %><%@
 page import="com.liferay.commerce.order.rule.web.internal.display.context.COREntryDisplayContext" %><%@
-page import="com.liferay.commerce.order.rule.web.internal.frontend.constants.COREntryDataSetDisplayNames" %><%@
-page import="com.liferay.commerce.order.rule.web.internal.frontend.taglib.servlet.taglib.constants.COREntryScreenNavigationEntryConstants" %><%@
+page import="com.liferay.commerce.order.rule.web.internal.entry.constants.COREntryDataSetDisplayNames" %><%@
+page import="com.liferay.commerce.order.rule.web.internal.entry.constants.COREntryScreenNavigationEntryConstants" %><%@
 page import="com.liferay.portal.kernel.bean.BeanParamUtil" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
 page import="com.liferay.portal.kernel.portlet.LiferayWindowState" %><%@
